backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.27.0
chromadb==1.3.7
ollama==0.6.1
python-multipart==0.0.6
watchdog==6.0.0
pypdf==6.4.2
python-docx==1.2.0
python-pptx==0.6.21
openpyxl==3.1.2
python-magic==0.4.27
aiofiles==23.2.1
pydantic==2.12.5
pydantic-settings==2.12.0